requirements.txt
# RXNRECer - Core Dependencies
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Protein Language Models
fair-esm>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0

# Bioinformatics
biopython>=1.80
rdkit>=2023.3.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
retrying>=1.3.0

# Optional: Parallel Processing (for large datasets)
pandarallel>=1.6.0

# Optional: Visualization (for evaluation)
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter support (for development)
jupyter>=1.0.0
ipykernel>=6.0.0 